+++ CHANGELOG.md
@@ -8,6 +8,8 @@ ChangeLog
 * Using php-cs-fixer for automated coding standards enforcement and fixing.
 * #660: principals could break html output.
 * #662: Fixed several bugs in the `share` request parser.
+* #665: Fix a bug in serialization of complex properties in the proppatch
+  request in the client.
 
 
 3.0.0-alpha1 (2015-05-19)